+++ R/colformat.R
@@ -1,10 +1,11 @@
 #' Format character columns as markdown text
 #'
 #' @param x A `flextable` object
+#' @param j Columns to be treated as markdown texts.
+#'   Selection can be done by the semantics of `dplyr::select()`.
 #' @param auto_color_link A color of the link texts.
 #' @param .from
-#'   Pandoc's `--from` argument (default: `'markdown'`).
-#'   Extentions can be enabled with `+`, i.e. `'markdown+emoji'`.
+#'   Pandoc's `--from` argument (default: `'markdown+autolink_bare_uris+emoji'`).
 #'
 #' @examples
 #' if (rmarkdown::pandoc_available()) {
@@ -16,14 +17,22 @@
 #' }
 #'
 #' @export
-colformat_md <- function(x, auto_color_link = 'blue', .from = 'markdown+emoji') {
+colformat_md <- function(
+  x, j = is.character,
+  auto_color_link = 'blue', .from = 'markdown+autolink_bare_uris+emoji'
+) {
   body <- x$body$dataset
-  j = names(body)[vapply(body, is.character, NA)]
-  if (length(j) == 0) return(x)
+  .j <- rlang::enexpr(j)
+  col <- names(tidyselect::eval_select(rlang::expr(c(!!.j)), body))
+
+  if (length(col) == 0) return(x)
+
   flextable::compose(
-    x, i = seq(nrow(body)), j = j,
+    x,
+    i = seq(nrow(body)),
+    j = col,
     value = parse_md(
-      unlist(body[j], use.names = FALSE), .from =  .from,
+      unlist(body[col], use.names = FALSE), .from = .from,
       auto_color_link = auto_color_link
     ),
     part = 'body'